# copilot-agent-and-mcp/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Book Management API
  description: A REST API for managing books, user favorites, and reading lists
  version: 1.0.0
  contact:
    name: API Support
    email: support@bookapi.com
  license:
    name: MIT
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: https://api.bookapp.com/v1
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /reading-lists:
    get:
      summary: Get user's reading list
      tags: [Reading Lists]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [want-to-read, currently-reading, finished]
      responses:
        '200':
          description: User's reading list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReadingListItem'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                  _links:
                    $ref: '#/components/schemas/PaginationLinks'

    post:
      summary: Add book to reading list
      tags: [Reading Lists]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookId, status]
              properties:
                bookId:
                  type: string
                status:
                  type: string
                  enum: [want-to-read, currently-reading, finished]
            example:
              bookId: "123"
              status: "want-to-read"
      responses:
        '201':
          description: Book added to reading list
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  book:
                    $ref: '#/components/schemas/ReadingListItem'
        '400':
          description: Invalid input or book already in list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reading-lists/{bookId}:
    put:
      summary: Update reading status (full replacement)
      tags: [Reading Lists]
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [want-to-read, currently-reading, finished]
      responses:
        '200':
          description: Reading status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Book not found in reading list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update reading status (partial update)
      tags: [Reading Lists]
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [want-to-read, currently-reading, finished]
      responses:
        '200':
          description: Reading status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

    delete:
      summary: Remove book from reading list
      tags: [Reading Lists]
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Book removed from reading list
        '404':
          description: Book not found in reading list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Book:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        year:
          type: integer
        description:
          type: string
        _links:
          type: object
          properties:
            self:
              type: object
              properties:
                href:
                  type: string
            favorite:
              type: object
              properties:
                href:
                  type: string
      example:
        id: "123"
        title: "The Great Gatsby"
        author: "F. Scott Fitzgerald"
        genre: "Classic Literature"
        year: 1925
        description: "A classic American novel..."
        _links:
          self:
            href: "/api/v1/books/123"
          favorite:
            href: "/api/v1/favorites"
    ReadingListItem:
      allOf:
        - $ref: '#/components/schemas/Book'
        - type: object
          properties:
            readingStatus:
              type: string
              enum: [want-to-read, currently-reading, finished]
            dateAdded:
              type: string
              format: date-time
            dateFinished:
              type: string
              format: date-time
              nullable: true
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
      example:
        page: 1
        limit: 20
        total: 150
        pages: 8
        hasNext: true
        hasPrev: false
    PaginationLinks:
      type: object
      properties:
        self:
          type: object
          properties:
            href:
              type: string
        first:
          type: object
          properties:
            href:
              type: string
        last:
          type: object
          properties:
            href:
              type: string
        next:
          type: object
          properties:
            href:
              type: string
        prev:
          type: object
          properties:
            href:
              type: string
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      example:
        message: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: string
      example:
        error:
          code: "VALIDATION_ERROR"
          message: "Invalid input provided"
          details: ["Username is required", "Password must be at least 6 characters"]
...